daemons/news_watch_daemon/pyproject.toml
[project]
name = "news-watch-daemon"
version = "0.1.0"
description = "Narrative-state engine for OpenClaw: scrapes headlines, synthesizes per-theme narratives, alerts on shifts."
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "feedparser>=6.0",
    "telethon>=1.40,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]

[project.scripts]
news-watch-daemon = "news_watch_daemon.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["news_watch_daemon*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
